src/main/java/kr/or/dgit/mybatis_dev/service/AddressService.java
package kr.or.dgit.mybatis_dev.service;

import java.util.List;

import org.apache.ibatis.session.RowBounds;
import org.apache.ibatis.session.SqlSession;

import kr.or.dgit.mybatis_dev.dao.AddressDao;
import kr.or.dgit.mybatis_dev.dao.AddressDaoImpl;
import kr.or.dgit.mybatis_dev.dto.Address;
import kr.or.dgit.mybatis_dev.util.MyBatisSqlSessionFactory;

public class AddressService {
	public List<Address> findAddressByAll(RowBounds rowBounds){
		try (SqlSession sqlsession = MyBatisSqlSessionFactory.getSqlSessionFactory().openSession()) {
			AddressDao dao = new AddressDaoImpl(sqlsession);
			return dao.selectAddressByAll(rowBounds);
		}
	}
	
	public List<Address> findAddressByAllWithAPI(RowBounds rowBounds){
		try (SqlSession sqlsession = MyBatisSqlSessionFactory.getSqlSessionFactory().openSession()) {
			AddressDao dao = new AddressDaoImpl(sqlsession);
			return dao.selectAddressByAllWithAPI(rowBounds);
		}
	}
}
